Move services
Took 4 minutes
