+++ DentalCore.Wpf/ViewModels/Factories/IViewModelFactory.cs
@@ -1,4 +1,4 @@
-﻿using DentalCore.Wpf.Services.Navigation;
+﻿using DentalCore.Wpf.Abstract;
 
 namespace DentalCore.Wpf.ViewModels.Factories;
