+++ DentalCore.Wpf/ViewModels/PatientCreateViewModel.cs
@@ -15,7 +15,6 @@ using DentalCore.Domain.Dto;
 using DentalCore.Domain.Services;
 using DentalCore.Wpf.Abstract;
 using DentalCore.Wpf.Commands;
-using DentalCore.Wpf.Services.Navigation;
 using DentalCore.Wpf.ViewModels.Components;
 using DentalCore.Wpf.ViewModels.Inners;
