+++ DentalCore.Wpf/ViewModels/PatientInfoViewModel.cs
@@ -10,7 +10,6 @@ using System.Windows.Input;
 using DentalCore.Domain.Abstract;
 using DentalCore.Wpf.Abstract;
 using DentalCore.Wpf.Commands;
-using DentalCore.Wpf.Services.Navigation;
 using DentalCore.Wpf.ViewModels.Inners;
 
 namespace DentalCore.Wpf.ViewModels;
