+++ DentalCore.Wpf/ViewModels/PatientsViewModel.cs
@@ -9,7 +9,6 @@ using DentalCore.Domain.Abstract;
 using DentalCore.Domain.Services;
 using DentalCore.Wpf.Abstract;
 using DentalCore.Wpf.Commands;
-using DentalCore.Wpf.Services.Navigation;
 using DentalCore.Wpf.ViewModels.Inners;
 
 namespace DentalCore.Wpf.ViewModels;
