+++ DentalCore.Wpf/ViewModels/VisitsViewModel.cs
@@ -9,7 +9,6 @@ using System.Windows.Input;
 using DentalCore.Domain.Abstract;
 using DentalCore.Wpf.Abstract;
 using DentalCore.Wpf.Commands;
-using DentalCore.Wpf.Services.Navigation;
 using DentalCore.Wpf.ViewModels.Inners;
 
 namespace DentalCore.Wpf.ViewModels;
