+++ DentalCore.Wpf/Views/Components/NavBar.xaml
@@ -4,7 +4,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 
-             xmlns:navigation="clr-namespace:DentalCore.Wpf.Services.Navigation"
+             xmlns:abstract="clr-namespace:DentalCore.Wpf.Abstract"
              mc:Ignorable="d"
              d:DesignHeight="300" d:DesignWidth="1000">
 
@@ -26,27 +26,27 @@
                          Style="{StaticResource NiceRadioButton}"
                          IsChecked="{Binding CurrentNavBarOption, Mode=OneWay,
                          Converter={StaticResource EqualsToBoolConverter},
-                         ConverterParameter={x:Static navigation:ViewType.Patients}}"
+                         ConverterParameter={x:Static abstract:ViewType.Patients}}"
                          Command="{Binding Path=Navigator.UpdateCurrentViewTypeCommand}"
-                         CommandParameter="{x:Static navigation:ViewType.Patients}"
+                         CommandParameter="{x:Static abstract:ViewType.Patients}"
                          Content="Пацієнти" />
 
             <RadioButton Grid.Column="1"
                          Style="{StaticResource NiceRadioButton}"
                          IsChecked="{Binding CurrentNavBarOption, Mode=OneWay,
                          Converter={StaticResource EqualsToBoolConverter},
-                         ConverterParameter={x:Static navigation:ViewType.Visits}}"
+                         ConverterParameter={x:Static abstract:ViewType.Visits}}"
                          Command="{Binding Path=Navigator.UpdateCurrentViewTypeCommand}"
-                         CommandParameter="{x:Static navigation:ViewType.Visits}"
+                         CommandParameter="{x:Static abstract:ViewType.Visits}"
                          Content="Візити" />
             
             <RadioButton Grid.Column="2"
                          Style="{StaticResource NiceRadioButton}"
                          IsChecked="{Binding CurrentNavBarOption, Mode=OneWay,
                          Converter={StaticResource EqualsToBoolConverter},
-                         ConverterParameter={x:Static navigation:ViewType.Procedures}}"
+                         ConverterParameter={x:Static abstract:ViewType.Procedures}}"
                          Command="{Binding Path=Navigator.UpdateCurrentViewTypeCommand}"
-                         CommandParameter="{x:Static navigation:ViewType.Procedures}"
+                         CommandParameter="{x:Static abstract:ViewType.Procedures}"
                          Content="Процедури" />
         </Grid>
     </Grid>
